* implemented task retry and incexec

+++ pandajedi/jediorder/TaskCommando.py
@@ -10,6 +10,8 @@ from pandajedi.jedicore.MsgWrapper import MsgWrapper
 from JediKnight import JediKnight
 from pandajedi.jediconfig import jedi_config
 from pandajedi.jedicore.JediTaskSpec import JediTaskSpec
+from pandajedi.jedirefine import RefinerUtils
+
 
 # logger
 from pandacommon.pandalogger.PandaLogger import PandaLogger
@@ -102,29 +104,63 @@ class TaskCommandoThread (WorkerThread):
                 # loop over all tasks
                 for jediTaskID,commandMap in taskList:
                     # make logger
-                    tmpLog = MsgWrapper(self.logger,'jediTaskID={0}'.format(jediTaskID))
+                    tmpLog = MsgWrapper(self.logger,' <jediTaskID={0}>'.format(jediTaskID))
                     commandStr = commandMap['command']
                     commentStr = commandMap['comment']
                     tmpLog.info('start for {0}'.format(commandStr))
                     tmpStat = Interaction.SC_SUCCEEDED
-                    # get active PandaIDs
-                    pandaIDs = self.taskBufferIF.getPandaIDsWithTask_JEDI(jediTaskID,True)
-                    if pandaIDs == None:
-                        tmpLog.error('failed to get PandaIDs for jediTaskID={0}'.format(jediTaskID))
-                        tmpStat = Interaction.SC_FAILED
-                    # kill jobs or update task
-                    if tmpStat == Interaction.SC_SUCCEEDED:
-                        if pandaIDs == []:
-                            # done since no active jobs
-                            tmpLog.info('completed the command')
-                            tmpTaskSpec = JediTaskSpec()
-                            tmpTaskSpec.jediTaskID = jediTaskID
-                            tmpTaskSpec.status = JediTaskSpec.terminateCommandStatusMap()[commandStr]['done']
-                            tmpRet = self.taskBufferIF.updateTask_JEDI(tmpTaskSpec,{'jediTaskID':jediTaskID})
-                        else:
-                            tmpLog.info('sending kill command')
-                            tmpRet = self.taskBufferIF.killJobs(pandaIDs,commentStr,50,True)
-                        tmpLog.info('done with {0}'.format(str(tmpRet)))
+                    if commandStr in ['kill','finish']:
+                        # get active PandaIDs to be killed
+                        pandaIDs = self.taskBufferIF.getPandaIDsWithTask_JEDI(jediTaskID,True)
+                        if pandaIDs == None:
+                            tmpLog.error('failed to get PandaIDs for jediTaskID={0}'.format(jediTaskID))
+                            tmpStat = Interaction.SC_FAILED
+                        # kill jobs or update task
+                        if tmpStat == Interaction.SC_SUCCEEDED:
+                            if pandaIDs == []:
+                                # done since no active jobs
+                                tmpLog.info('completed the command')
+                                tmpTaskSpec = JediTaskSpec()
+                                tmpTaskSpec.jediTaskID = jediTaskID
+                                tmpTaskSpec.forceUpdate('oldStatus')
+                                tmpTaskSpec.status = JediTaskSpec.commandStatusMap()[commandStr]['done']
+                                tmpRet = self.taskBufferIF.updateTask_JEDI(tmpTaskSpec,{'jediTaskID':jediTaskID})
+                            else:
+                                tmpLog.info('sending kill command')
+                                tmpRet = self.taskBufferIF.killJobs(pandaIDs,commentStr,50,True)
+                            tmpLog.info('done with {0}'.format(str(tmpRet)))
+                    elif commandStr in ['retry','incexec']:
+                        # change task params for incexec
+                        if commandStr == 'incexec':
+                            try:
+                                # read task params
+                                taskParam = self.taskBufferIF.getTaskParamsWithID_JEDI(jediTaskID)
+                                taskParamMap = RefinerUtils.decodeJSON(taskParam)
+                                # convert new params
+                                newParamMap = RefinerUtils.decodeJSON(commentStr)
+                                # change params
+                                for newKey,newVal in newParamMap.iteritems():
+                                    if newVal == None:
+                                        # delete 
+                                        del taskParamMap[newKey]
+                                    else:
+                                        # change
+                                        taskParamMap[newKey] = newVal
+                                # encode new param
+                                strTaskParams = RefinerUtils.encodeJSON(taskParamMap)
+                                tmpRet = self.taskBufferIF.updateTaskParams_JEDI(jediTaskID,strTaskParams)
+                                if tmpRet != True:
+                                    tmpLog.error('failed to update task params')
+                                    continue
+                            except:
+                                errtype,errvalue = sys.exc_info()[:2]
+                                tmpLog.error('failed to change task params with {0}:{1}'.format(errtype.__name__,errvalue))
+                                continue
+                        # retry failed files
+                        tmpRet = self.taskBufferIF.retryTask_JEDI(jediTaskID,commandStr)
+                        tmpLog.info('done with {0}'.format(tmpRet))
+                    else:
+                        tmpLog.error('unknown command')
             except:
                 errtype,errvalue = sys.exc_info()[:2]
                 logger.error('{0} failed in runImpl() with {1}:{2}'.format(self.__class__.__name__,errtype.__name__,errvalue))
